+++ raytracer/objects/Sphere.cpp
@@ -1,7 +1,7 @@
 #include "Sphere.h"
 #include "Ray.h"
 
-float Sphere::intersect(const Ray* ray) const {
+double Sphere::intersect(const Ray* ray) const {
     // a * t ^ 2 + b * t + c = 0
     // t ^ 2 * (P1 * P1) + 2 * t * P1 * (P0 - C) + (P0 - C) * (P0 - C) - r ^ 2 = 0
     // a = P1 * P1
@@ -14,25 +14,25 @@ float Sphere::intersect(const Ray* ray) const {
     vec4 rayDir = invM * ray->dir;
 
     glm::vec4 diff = rayPos - position;
-    float a = glm::dot(rayDir, rayDir);
-    float b = 2 * glm::dot(rayDir, diff);
-    float c = glm::dot(diff, diff) - powf(radius, 2);
-    float D = powf(b, 2) - 4 * a * c;
+    double a = glm::dot(rayDir, rayDir);
+    double b = 2 * glm::dot(rayDir, diff);
+    double c = glm::dot(diff, diff) - powf(radius, 2);
+    double D = powf(b, 2) - 4 * a * c;
     if (D < 0) {
         return 0;
     }
     if (D == 0) {
         return -b / (2 * a);
     }
-    float sqD = sqrtf(D);
-    float t1 = -(b + sqD) / (2 * a);
-    float t2 = (sqD - b) / (2 * a);
+    double sqD = sqrtf(D);
+    double t1 = -(b + sqD) / (2 * a);
+    double t2 = (sqD - b) / (2 * a);
 
     if (t1 < 0 && t2 < 0) {
         return 0;
     }
 
-    float t;
+    double t;
     if (t1 < 0) {
         t = t2;
     } else if (t2 < 0) {
